Add a temporary shader program to debug render issue in OS X.
Add CurveEdit but currently used to debug.

+++ include/blendint/gui/abstract-view.hpp
@@ -602,7 +602,7 @@ private:
 
   bool destroying_;
 
-  int view_type_;
+  short view_type_;
   
   int reference_count_;
